Add files via upload
Shattajit loves travelling. One day he travels to a weird country. People of that land are a bit unconscious. They consider 3 and 7 unlucky. That’s why they do not use these two digits. So they count numbers like this: 0, 1, 2, 4, 5, 6, 8, 9, 10, 11, 12, 14, 15, 16, 18, 19, 20….

He went to a shop in this land to buy some things. But the cost is written in their weird number system without digits 3 & 7. Now your task is to help him convert and sum these wired numbers to the traditional numbers we use.